Added tolerance percentages
Each object is now generated with a random tolerance to transmutation, identification, and seeding. These are influenced by its traits and each have a variability by up to 50%, which allows for percentages over 100 and under 0. 

This will allow the player to calculate with rough accuracy whether or not a certain action will work on a certain object.